Add property DefaultValueRaw
The "magic" behavior of DefaultValue for string types is
very helpful for most scenarios, but it makes one scenario
particularly difficult: if you want to use a string field/const
as default value for a string autobindable property.
See test case "Create_With_DefaultValue_As_String_Code"

This change adds an extra property DefaultValueRaw
which is always written as-is in the generated code, no matter
of what type the property has.